Moved testing harness to mocha Updated standard config Added thunk and logger middleware to redux Added test stubs
