+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "test": "tape tests/**/*Test.js",
+    "test": "mocha",
     "build-map-data": "node scripts/exportKMLToAnnotations.js",
     "build-android": "cd android && ./gradlew assembleRelease && cd ../",
     "install-release-android": "cd android && ./gradlew installRelease && cd ../"
@@ -15,13 +15,35 @@
     "react-native-mapbox-gl": "^4.1.0",
     "react-native-simple-store": "^0.1.0",
     "react-redux": "^4.4.5",
-    "redux": "^3.5.2"
+    "redux": "^3.5.2",
+    "redux-logger": "^2.6.1",
+    "redux-thunk": "^2.1.0"
   },
   "devDependencies": {
+    "babel-eslint": "^6.0.4",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-register": "^6.9.0",
+    "chai": "^3.5.0",
+    "eslint": "^2.12.0",
     "jsdom": "^8.3.1",
+    "mocha": "^2.5.3",
     "node-uuid": "^1.4.7",
     "standard": "^5.4.1",
-    "tape": "^4.4.0",
     "togeojson": "^0.13.0"
+  },
+  "babel": {
+    "presets": [
+      "es2015",
+      "react"
+    ]
+  },
+  "standard": {
+    "parser": "babel-eslint",
+    "globals": [
+      "describe",
+      "it",
+      "expect"
+    ]
   }
 }
