+++ src/RootViewContainer.js
@@ -1,13 +1,23 @@
 import React from 'react-native'
-import {createStore} from 'redux'
+import {createStore, applyMiddleware} from 'redux'
 import {Provider} from 'react-redux'
+import thunk from 'redux-thunk'
 import RootView from './RootView'
 import reducers from './reducers'
+import createLogger from 'redux-logger'
+
+const logger = createLogger()
 
 export default React.createClass({
   getInitialState () {
     return {
-      store: createStore(reducers)
+      store: createStore(
+        reducers,
+        applyMiddleware(
+          thunk,
+          logger
+        )
+      )
     }
   },
 
@@ -17,5 +27,5 @@ export default React.createClass({
         <RootView/>
       </Provider>
     )
-  },
+  }
 })
