+++ src/events/__tests__/eventsReducer.spec.js
@@ -1,0 +1,29 @@
+import eventsReducer from '../eventsReducer'
+
+describe('eventsReducer', () => {
+  it('should set initial state', () => {
+    // event detail visible
+    // events
+    // filter criteria
+    // static error message
+    // should show error message
+    // is loading
+
+  })
+
+  it('should handle data:load:complete', () => {
+    
+  })
+
+  it('should handle data:load:error', () => {
+    
+  })
+
+  it('should handle events:filter', () => {
+    
+  })
+
+  it('should handle events:sort', () => {
+    
+  });
+})
