+++ src/events/eventsReducer.js
@@ -1,9 +1,7 @@
-export default eventsReducer(state = defaultState(), action = {}) {
-  switch(action.type) {
-    default return state
+export default function eventsReducer (state = defaultState(), action = {}) {
+  switch (action.type) {
+    default: return state
   }
 }
 
-function defaultState() {
-  
-}
+function defaultState () {}
