+++ src/map/__tests__/mapReducer.spec.js
@@ -1,0 +1,19 @@
+import mapReducer from '../mapReducer'
+
+describe('mapReducer', () => {
+  it('should set its initial state', () => {
+    
+  })
+
+  it('should handle map:filter:toggle', () => {
+    
+  })
+
+  it('should handle data:load:complete', () => {
+    
+  })
+
+  it('should handle data:load:error', () => {
+    
+  })
+})
