+++ src/tabs/__tests__/tabsReducer.spec.js
@@ -1,0 +1,11 @@
+import tabsReducer from '../tabsReducer'
+
+describe('tabsReducer', () => {
+  it('should set its initial state', () => {
+    // list of routes, activeTabName
+  })
+
+  it('should handle tabs:active:change', () => {
+    // comes from setActiveTabName being called from the root view
+  })
+})
